feat: rename files in `vite-config` to match kebab-case
