+++ packages/vite-config/lib/with-page-config.ts
